refactor: split ajax logic into modules
