+++ src/main/webapp/WEB-INF/tags/footer.jsp
@@ -1,7 +1,8 @@
 </main> <footer class="footer mt-auto no-print">
         <div class="container text-center py-4">
-            <p class="mb-1">&copy; 2025 Livraria Mil Páginas - Todos os direitos reservados.</p>
-            <p class="small text-muted">Projeto Acadêmico de E-commerce</p>
+    <script type="module" src="${pageContext.request.contextPath}/js/main.js"></script>
+            <p class="mb-1">&copy; 2025 Livraria Mil PÃ¡ginas - Todos os direitos reservados.</p>
+            <p class="small text-muted">Projeto AcadÃªmico de E-commerce</p>
         </div>
     </footer>
